requirements.txt
discord.py
python-dotenv
httpx
pytest-asyncio
pytest
openai
duckduckgo-search
yt-dlp
websockets
